Retrospectively change minimum package version to 0.0.1
Recent versions of Julia deliberately check for a package registered
as 0.0.0 and throw an error. This is fine for the General registry but
we have many packages that either are or were once registered as 0.0.0.
This leads to errors when you perform package operations, even when not
on those particular packages (it manifests as a resolver error, and the
resolver runs on everything).

This has the potential to be disruptive, and I've not gone through and
edited the Project.toml files of the affected packages to be consistent,
but hopefully LocalRegistry will note this when we go to register new
versions of these packages.